mavlink_parameters: refactor to one work queue
To make everything simpler it makes more sense to have one work queue
instead of one for get and one for set. This way we just work through
the items as they come in instead of an (unexpected) priorization of get
over set.

Also, this means that we can get rid of some code duplication, and other
uglyness.